+++ asdf/_tests/_helpers.py
@@ -453,9 +453,7 @@ def _assert_extension_type_correctness(extension, extension_type, resolver):
 
     # Currently ExtensionType sets a default version of 1.0.0,
     # but we want to encourage an explicit version on the subclass.
-    assert "version" in extension_type.__dict__, "{} must set the 'version' class attribute".format(
-        extension_type.__name__,
-    )
+    assert "version" in extension_type.__dict__, f"{extension_type.__name__} must set the 'version' class attribute"
 
     # check the default version
     types_to_check = [extension_type]
@@ -490,7 +488,7 @@ def _assert_extension_type_correctness(extension, extension_type, resolver):
             try:
                 with generic_io.get_file(schema_location) as f:
                     yaml.safe_load(f.read())
-            except Exception as err:  # noqa: BLE001
+            except Exception as err:
                 msg = (
                     f"{extension_type.__name__} supports tag, {check_type.yaml_tag}, "
                     f"which resolves to schema at {schema_location}, but "