+++ asdf/_tests/tags/core/tests/test_history.py
@@ -3,14 +3,13 @@ import fractions
 import os
 
 import pytest
-from jsonschema import ValidationError
 
 import asdf
 from asdf import _types as types
 from asdf import util
 from asdf._tests import _helpers as helpers
 from asdf._tests._helpers import assert_no_warnings, yaml_to_asdf
-from asdf.exceptions import AsdfDeprecationWarning, AsdfWarning
+from asdf.exceptions import AsdfDeprecationWarning, AsdfWarning, ValidationError
 from asdf.tags.core import HistoryEntry
 
 SCHEMA_PATH = os.path.join(os.path.dirname(helpers.__file__), "data")