+++ asdf/tags/core/ndarray.py
@@ -2,10 +2,10 @@ import mmap
 import sys
 
 import numpy as np
-from jsonschema import ValidationError
 from numpy import ma
 
 from asdf import _types, util
+from asdf._jsonschema import ValidationError
 
 _datatype_names = {
     "int8": "i1",
@@ -87,7 +87,7 @@ def asdf_datatype_to_numpy_dtype(datatype, byteorder=None):
 
             else:
                 msg = "Error parsing asdf datatype"
-                raise RuntimeError(msg)  # noqa: TRY004
+                raise RuntimeError(msg)
 
         return np.dtype(datatype_list)