+++ asdf/util.py
@@ -371,8 +371,8 @@ def minversion(module, version, inclusive=True):
         except ImportError:
             return False
     else:
-        msg = f"module argument must be an actual imported module, or the import name of the module; got {repr(module)}"
-        raise ValueError(msg)  # noqa: TRY004
+        msg = f"module argument must be an actual imported module, or the import name of the module; got {module!r}"
+        raise ValueError(msg)
 
     if module_version is None:
         try: